+++ src/Codec/FFmpeg/Decode.hs
@@ -10,7 +10,7 @@ import Codec.FFmpeg.Types
 import Control.Applicative
 import Control.Arrow (first)
 import Control.Monad (when, void)
-import Control.Monad.Error.Class
+import Control.Monad.Except
 import Control.Monad.IO.Class
 import Control.Monad.Trans.Maybe
 import Foreign.C.String